added return Double[]
Also cleaned up the code, added some comments.

Next step is to figure out a way to deliver the coordinates in a simpler way, where it is clearer which one is which.

I'll also remove the main method in a later version when it's all done and I don't need to test it anymore